feat: Move Routes to component feat: Add Global Modal in order to not open./close at each url change
