+++ packages/cozy-harvest-lib/src/index.js
@@ -7,7 +7,7 @@ export {
 } from './components/hoc/withKonnectorModal'
 export { default as TriggerManager } from './components/TriggerManager'
 export { default as TriggerLauncher } from './components/TriggerLauncher'
+export { default as Routes } from './components/Routes'
 
 export { KonnectorJobError } from './helpers/konnectors'
 export { handleOAuthResponse } from './helpers/oauth'
-export { default as Routes } from './Routes'
